backend/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.34.0
httpx==0.27.2
qdrant-client==1.12.1
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
tenacity==9.0.0
SQLAlchemy==2.0.30
psycopg[binary]==3.1.19
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.2.1
python-multipart==0.0.9
beautifulsoup4==4.12.3
lxml==5.3.0
Pillow==10.4.0

